moved the chat to different endpoint and updated the home page to include card with a link that takes you to the chat room
